+++ NEWS.md
@@ -1,5 +1,8 @@
 # Changes in version 0.2-5.1 (xx/xx/2019)
 
+## Bugfixes
+* predict for Recommender RANDOM now uses the correct user ids in the prediction (reported by aliko-str). 
+
 # Changes in version 0.2-5 (08/27/2019)
 
 ## Changes